完善user中新加的接口，favorite和feed部分debug

+++ service/favorite/favorite.go
@@ -4,28 +4,28 @@ import (
 	"Douyin/common"
 	"Douyin/repository"
 	"Douyin/service/user"
+	"errors"
 )
 
 func FavoriteAction(userId uint, VideoId uint, actionType uint) (err error) {
-
+	if userId == 0 {
+		return errors.New("user not logged in")
+	}
 	isFavorite := repository.IsBFavoriteA(VideoId, userId)
-
-	// 点赞操作
-	if actionType == 1 {
-		if !isFavorite {
-			err := repository.AddFavorite(VideoId, userId)
-			if err != nil {
-				return err
-			}
+	if actionType == 1 && !isFavorite {
+		// 点赞操作
+		err := repository.AddFavorite(VideoId, userId)
+		if err != nil {
+			return err
 		}
+	} else if actionType == 0 && isFavorite {
 		// 取消点赞操作
-	} else {
-		if isFavorite {
-			err := repository.DeleteFavorite(VideoId, userId)
-			if err != nil {
-				return err
-			}
+		err := repository.DeleteFavorite(VideoId, userId)
+		if err != nil {
+			return err
 		}
+	} else {
+		return errors.New("invalid action")
 	}
 	return nil
 }
@@ -46,7 +46,7 @@ func GetFavoriteList(userNewId uint, userHostId uint) []common.Video {
 		favoriteUser, _ := user.GetUserInfo(x.AuthorId, userHostId)
 		tmp.Author = favoriteUser
 
-		tmp.FavoriteCount = repository.GetFavoriteCountById(tmp.Id)
+		tmp.FavoriteCount = repository.GetFavoritedCountByVideoId(tmp.Id)
 		tmp.CommentCount = repository.GetCommentCountById(tmp.Id)
 		tmp.IsFavorite = repository.IsBFavoriteA(tmp.Id, userHostId)
 
@@ -58,7 +58,7 @@ func GetFavoriteList(userNewId uint, userHostId uint) []common.Video {
 
 func FavoriteGet(userId uint) []repository.Video {
 
-	var favoriteList = repository.GetFavoriteById(userId)
+	var favoriteList = repository.GetFavoriteByUserId(userId)
 
 	videoList := make([]repository.Video, 0)
